# docs/swagger.yaml
definitions:
  v1.AssignContentItemTagsRequest:
    properties:
      tagIds:
        items:
          type: integer
        type: array
    required:
    - tagIds
    type: object
...
  v1.CreateTagRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  v1.ListTagsResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/v1.ListTagsResponseData'
      message:
        type: string
    type: object
  v1.ListTagsResponseData:
    properties:
      items:
        items:
          $ref: '#/definitions/v1.TagResponseData'
        type: array
      total:
        type: integer
    type: object
...
  v1.RenameTagRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  v1.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  v1.TagResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/v1.TagResponseData'
      message:
        type: string
    type: object
  v1.TagResponseData:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Nunu Example API
  version: 1.0.0
paths:
...
  /content-items/{id}/tags:
    delete:
      consumes:
      - application/json
      description: 从单个内容条目移除多个标签。
      parameters:
      - description: content item id
        in: path
        name: id
        required: true
        type: integer
      - description: tag ids
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.AssignContentItemTagsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Response'
      security:
      - Bearer: []
      summary: 移除内容条目标签
      tags:
      - 标签模块
    put:
      consumes:
      - application/json
      description: 给单个内容条目添加多个已有标签，不即时新建标签。
      parameters:
      - description: content item id
        in: path
        name: id
        required: true
        type: integer
      - description: tag ids
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.AssignContentItemTagsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Response'
      security:
      - Bearer: []
      summary: 给内容条目添加标签
      tags:
      - 标签模块
...
  /tags:
    get:
      description: 返回全部标签。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListTagsResponse'
      security:
      - Bearer: []
      summary: 查询标签列表
      tags:
      - 标签模块
    post:
      consumes:
      - application/json
      description: 创建用于组织内容条目的标签，名称唯一。
      parameters:
      - description: tag
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TagResponse'
      security:
      - Bearer: []
      summary: 创建标签
      tags:
      - 标签模块
  /tags/{id}:
    delete:
      description: 删除标签本身及其与内容条目的关联，不删除内容条目。
      parameters:
      - description: tag id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Response'
      security:
      - Bearer: []
      summary: 删除标签
      tags:
      - 标签模块
    put:
      consumes:
      - application/json
      description: 修改已有标签名称。
      parameters:
      - description: tag id
        in: path
        name: id
        required: true
        type: integer
      - description: tag
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.RenameTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TagResponse'
      security:
      - Bearer: []
      summary: 重命名标签
      tags:
      - 标签模块
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"